+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+- `MiniAdvvertCardView` added new component, info colored button with Right chevron icon
 
 ## [7.1.3] - 2023-01-10Z
 - `SummaryRowView` suffixAccessibilityLabel suffixLabel parameter added following DAC report